requirements.txt
vpandas
numpy
scikit-learn
joblib
streamlit